placed rng feature into its own folder and removed test class
