+++ Assets/Scripts/Features/RNG.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 39468e08c657dcf45990ffef26a98bda
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
