+++ Assets/Scripts/Features/RNG/IChanceScore.cs
