--- Assets/Scripts/TestClass.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: a30c7e96204c48d43be813575d98e90f
